+++ ansible/hosts.ini
@@ -1,3 +1,3 @@
 [app_servers]
-app1 ansible_host=18.185.54.106
-app2 ansible_host=52.58.89.62
+app1 ansible_host=18.184.11.13
+app2 ansible_host=3.73.38.3
