--- myproject/requirements_dev.txt
@@ -1,1 +1,0 @@
-Flask-Testing
