+++ src/requirements_dev.txt
@@ -1,0 +1,2 @@
+Flask-Testing
+bandit
